backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
python-dotenv==1.0.1
deepgram-sdk==3.7.7
anthropic==0.34.2
httpx==0.27.2
numpy==2.1.1
python-multipart==0.0.12
